requirements.txt
asgiref==3.8.1
black==24.4.2
click==8.1.7
Django==5.0.7
django-allauth==64.0.0
django-cors-headers==4.4.0
django-debug-toolbar==4.4.6
django-extensions==3.2.3
django-filter==24.2
djangorestframework==3.15.2
gunicorn==22.0.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
python-dotenv==1.0.1
pytz==2024.1
sqlparse==0.5.1
whitenoise==6.7.0
